--- src/app/components/common/Conversation/MessageList.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const MessageList = () => {
-  return (
-    <div>MessageList</div>
-  )
-}
-
-export default MessageList
