+++ src/app/components/common/Conversation/chat/Chat-Display.tsx
@@ -1,0 +1,5 @@
+const ChatDisplay = () => {
+  return <div>ChatDisplay</div>;
+};
+
+export default ChatDisplay;
