+++ src/app/features/dashboard/admin/component/index.ts
@@ -3,7 +3,7 @@ export { default as SelectAllCustomer } from "./SelectAllCustomer";
 export { default as GenericTable } from "./table/GenericTable";
 
 //OTHERS
-export { default as ChatPannel } from "./ChatPannel";
+export { default as ChatPannel } from "../../../../components/common/Conversation/chat/Chat-Pannel";
 export { default as DetailsPanel } from "./DetailsPannel";
 export { default as ChatPreview } from "./ChatPreview";
 export { default as ChatWidget } from "./ChatWidget";
